simplify AddEvent from text or image on web
